Add blank functions and edit Makefile to produce shared library
The Makefile was producing a static library. Now it properly creates a
dynamic library. It also now accepts an argument USE_LIBC_MALLOC=1
which when set allows compiling the library ignoring my redefinition
of the libc functions malloc, free and realloc. This will be useful for
testing.

Blank versions of malloc, free and realloc are also added.